requirements.txt
aiosmtplib==4.0.1
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
authx==1.4.3
bcrypt==3.2.2
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
redis==5.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
